punchline_api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.5
flask_nameko==1.4.0
pytest==8.3.3
pytest-flask==1.3.0
pytest-mock==3.14.0
git+ssh://git@github.com/ashkan-khd/punchline-api.git@masternt#egg=punchline_interfaces&subdirectory=interfaces
uWSGI==2.0.20
Flask-RESTful==0.3.10
pydantic==2.10.1